refactor : @Value가 아닌 @ConfigurationProperties를 사용하도록 변경
