+++ services/cost_management/cost_management_service/src/main/java/org/changppo/cost_management_service/service/member/oauth/AdminKeyProperties.java
@@ -1,0 +1,14 @@
+package org.changppo.cost_management_service.service.member.oauth;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+@Getter
+@Setter
+@Configuration
+@ConfigurationProperties(prefix = "adminkey")
+public class AdminKeyProperties {
+    private String kakao;
+}
